api/requirements.txt
tensorflow==2.20.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.19.1
matplotlib
numpy
